pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raven-security"
version = "0.1.0"
description = "AI-powered offensive security research platform"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.10"
authors = [
    {name = "RAVEN Team"},
]
keywords = ["security", "binary-analysis", "exploit", "reverse-engineering", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Disassemblers",
]
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "pyyaml>=6.0.1",
    "pyelftools>=0.30",
    "pefile>=2023.2.7",
    "capstone>=5.0.1",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
llm = [
    "openai>=1.12.0",
    "anthropic>=0.18.0",
]
full = [
    "raven-security[llm]",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "keystone-engine>=0.9.2",
    "unicorn>=2.0.1",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.3.0",
    "mypy>=1.7.0",
]

[project.scripts]
raven = "raven.cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/raven"
Documentation = "https://github.com/yourusername/raven/tree/main/docs"
Repository = "https://github.com/yourusername/raven"
Issues = "https://github.com/yourusername/raven/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
